Modificado governor

+++ PID_controler/pid.h
@@ -5,7 +5,7 @@
 #define FREQ_FILE "/sys/devices/system/cpu/cpu0/cpufreq/scaling_available_frequencies"
 
 // Choose the correct frequency
-int which_freq(int freq);
+int which_freq(float freq);
 
 // Update the temperature and return the choosen frequency 
 int update_temp(int error);
